+++ rdns_reaper/__init__.py
@@ -1,3 +1,3 @@
 from .rdns_reaper import rdns_reaper
 
-__version__ = "0.0.11"
+__version__ = "0.0.12"