+++ rdns_reaper/rdns_reaper.py
@@ -1,3 +1,11 @@
+"""
+Reverse DNS Lookup Engine.
+
+This module allows other programs to load in multipe IPv4 or IPv6 addresses and then
+perform a serial or parallel resolving of all addresses.  Filtering options exist
+and a disk based cache is supported to decrease runtime for subsequent uses.
+"""
+
 import copy
 import socket
 import concurrent.futures
@@ -43,9 +51,34 @@ IPV6_RESERVED_NETWORK_LIST = [
 
 
 class rdns_reaper:
-    def __init__(self, *args, **kwargs):
-        """Initialize class and take in user options."""
-        self._dns_dict = dict()
+    """Reverse DNS Lookup Engine."""
+
+    _filter = None
+
+    def __init__(self, **kwargs):
+        """Initialize class and take in user options.
+
+        Keyword Arguments:
+            allow_reserved_networks (bool, optional): if True disable automatic filtering of
+                reserved networks, must be set to True if checking of any reserved networks
+                is desired.  Can then be supplemented with a custom filter
+            concurrent (int, default = 5): number of concurrent resolver threads
+            limit_to_rfc1918 (bool, optional): limit resolve to IPv4 RFC1918 only
+            filter (str, list of strs, IPSet, optional): filter data
+                can be a string containing an IP Address, list of strings, or an IPSet object
+            filtermode ("block" | "allow", optional): sets filter mode to block list or allow list
+                defaults to block list if not specified
+            filename (str, optional): path and filename for disk based cache in YAML format
+            filemode ("r" | "w"), required if filename set): read only or read-write mode for
+                disk cache, if set to write, resolver may try to update the cache periodically
+                the savefile() function can be called without arguments to force an update
+            unresolvable (str, optional): string to set for each entry if resolving fails
+                None is the default, all other options prevent subsequent lookups if disk based
+                cache is in use
+
+
+        """
+        self._dns_dict = {}
         self._resolver_ip = None
 
         """Check for RFC1918 filtering"""
@@ -53,6 +86,23 @@ class rdns_reaper:
             self.limit_to_rfc1918(kwargs["limit_to_rfc1918"])
         else:
             self._limit_to_rfc1918 = False
+
+        """Check for custom filtering"""
+        if kwargs.get("filter") is not None:
+            if kwargs.get("filtermode") is not None:
+                self.set_filter(kwargs["filter"], mode=kwargs["filtermode"])
+            else:
+                self.set_filter(kwargs["filter"])
+        else:
+            self._filter = None
+            self._filter_mode = None
+
+        """Allow reserved network check"""
+
+        if kwargs.get("allow_reserved_networks") is True:
+            self._allow_reserved_networks = True
+        else:
+            self._allow_reserved_networks = False
 
         """Process parallel lookup concurrency"""
         try:
@@ -65,7 +115,7 @@ class rdns_reaper:
 
         """Determine how to mark unresolvable entries"""
         try:
-            if type(kwargs["unresolvable"]) is str:
+            if isinstance(kwargs["unresolvable"], str):
                 self._unresolvable = kwargs["unresolvable"]
             else:
                 raise TypeError
@@ -73,7 +123,7 @@ class rdns_reaper:
             self._unresolvable = None
 
         try:
-            if type(kwargs["filemode"]) is str:
+            if isinstance(kwargs["filemode"], str):
                 self._filemode = kwargs["filemode"]
             else:
                 raise TypeError
@@ -81,16 +131,14 @@ class rdns_reaper:
             self._filemode = None
 
         try:
-            if type(kwargs["filename"]) is str:
+            if isinstance(kwargs["filename"], str):
                 self._filename = kwargs["filename"]
             else:
                 raise TypeError
         except KeyError:
             self._filename = None
 
-        if (self._filename is not None) and (
-            (self._filemode == "r") or (self._filemode == "w")
-        ):
+        if (self._filename is not None) and ((self._filemode in "r", "w")):
             try:
                 self.loadfile(self._filename)
             except FileNotFoundError:
@@ -101,25 +149,33 @@ class rdns_reaper:
         self_copy = copy.deepcopy(self)
         if type(new) is rdns_reaper:
             self_copy._dns_dict.update(new._dns_dict)
-        elif type(new) is str:
+        elif isinstance(new, str):
             self_copy.add_ip(new)
-        elif type(new) is set:
+        elif isinstance(new, set):
             self_copy.add_ip_list(new)
-        elif type(new) is list:
+        elif isinstance(new, list):
             self_copy.add_ip_list(new)
         else:
             raise TypeError
         return self_copy
 
-    def __contains__(self, ip):
-        """Allow for checking of IP in the instance."""
-        if ip in self._dns_dict.keys():
+    def __contains__(self, ip_address):
+        """Allow for checking of IP in the instance.
+
+        Args:
+            ip_address (str): A string containing an IP address to check
+
+        Returns:
+            bool: True if found, false if not found
+
+        """
+        if ip_address in self._dns_dict.keys():
             return True
         return False
 
-    def __delitem__(self, ip):
+    def __delitem__(self, ip_address):
         """Remove an item if del() is called."""
-        self.remove_ip(ip)
+        self.remove_ip(ip_address)
 
     def __enter__(self):
         """Return self for use in with blocks."""
@@ -130,13 +186,13 @@ class rdns_reaper:
         if (self._filename is not None) and (self._filemode == "w"):
             self.savefile(self._filename)
 
-    def __getitem__(self, ip):
+    def __getitem__(self, ip_address):
         """Return IP address from internal dictionary, false if not found."""
         try:
-            IPAddress(ip)
-            return self._dns_dict[ip]
-        except AddrFormatError:
-            raise TypeError
+            IPAddress(ip_address)
+            return self._dns_dict[ip_address]
+        except AddrFormatError as error_case:
+            raise TypeError from error_case
         except KeyError:
             return False
 
@@ -144,11 +200,11 @@ class rdns_reaper:
         """Add two instances together and return."""
         if type(new) is rdns_reaper:
             self._dns_dict.update(new._dns_dict)
-        elif type(new) is str:
+        elif isinstance(new, str):
             self.add_ip(new)
-        elif type(new) is set:
+        elif isinstance(new, set):
             self.add_ip_list(new)
-        elif type(new) is list:
+        elif isinstance(new, list):
             self.add_ip_list(new)
         else:
             raise TypeError
@@ -173,28 +229,64 @@ class rdns_reaper:
         except socket.herror:
             self._dns_dict[ip] = self._unresolvable
 
-    def add_ip(self, ip, hostname=None):
-        """Add an IP to the list with option hostname, skip if exists."""
-        if ip in self._dns_dict.keys():
+    def add_ip(self, ip_address, hostname=None):
+        """Add an IP to the list with option hostname, skip if exists.
+
+        Args:
+            ip_address (str): A string containing an IP address in v4 or v6 format
+            hostname (str): Optional string with a hostname
+                Without this argument, the Hostname will be set to None and will be eligible
+                for automatic lookup when the resolver functions are called
+                The hostname string should be a FQDN or similar text that would be returned
+                by a true DNS based reverseloopkup.
+
+        Returns:
+            True if already in dictionary, otherwise returns the key string back on success.
+
+        """
+        if ip_address in self._dns_dict.keys():
             return True
 
         try:
-            IPAddress(ip)
+            IPAddress(ip_address)
             if hostname is None:
-                self._dns_dict.update({ip: None})
+                self._dns_dict.update({ip_address: None})
             else:
-                self._dns_dict.update({ip: hostname})
-        except AddrFormatError:
-            raise TypeError
+                self._dns_dict.update({ip_address: hostname})
+            return ip_address
+        except AddrFormatError as error_case:
+            raise TypeError from error_case
 
     def add_ip_list(self, ip_list):
-        """Add all new IP's from a list."""
-        if type(ip_list) is set:
+        """Add all new IP's from a list.
+
+        Args:
+            ip_address (list): A list of strings each containing an IP address in v4 or v6 format
+        """
+        if isinstance(ip_list, set):
             ip_list = list(ip_list)
-        if type(ip_list) is not list:
+        if not isinstance(ip_list, list):
             raise TypeError
         for ip in ip_list:
             self.add_ip(ip)
+
+    def allow_reserved_networks(self, option):
+        """Allow users to enable/disable automatic filtering of reserved networks.
+
+        If a user wants to check reserved network IPs (loopbacks, link local, multicast, etc.)
+        they must set this option to True.  Users may manually filter some of the reserved
+        networks with the filter() option.  This option *must* be set to True to resolve any
+        reserved networks.  Using an allow filter with this option set to False will not resolve
+        any reserved networks.
+
+        Args:
+            option (bool): Set to True to disable automatic filtering of reserved networks
+
+        """
+        if not isinstance(option, bool):
+            raise TypeError
+
+        self._allow_reserved_networks = option
 
     def clear_all_hostnames(self):
         """Clear all the hostnames from existing entries."""
@@ -202,12 +294,17 @@ class rdns_reaper:
         self._dns_dict = new_ip_dict
 
     def clearname(self, ip):
-        """Clear a specific IP's hostname."""
+        """Clear a specific IP's hostname.
+
+        Args:
+            ip (str): A string containing an IP address
+        """
         try:
             ip = IPAddress(ip)
             self._dns_dict[ip] = None
-        except AddrFormatError:
-            raise TypeError
+            return True
+        except AddrFormatError as error_case:
+            raise TypeError from error_case
         except KeyError:
             return False
 
@@ -219,51 +316,108 @@ class rdns_reaper:
         """Return the internal dictionary to the calling function."""
         return self._dns_dict
 
+    def get_filter(self):
+        """Return current filter status.
+
+        Returns:
+            tuple - (filter, filter_mode)
+                filter is an IPSet object containing all networks that are part of the filter
+                filter_mode is a string of "block" or "allow" to indicate the filtering mode
+                    as a block list or allow list
+        """
+        try:
+            return (self._filter, self._filter_mode)
+        except AttributeError:
+            return None
+
+    def get_options(self):
+        """Return info about the various options set by the user."""
+        options_dict = {
+            "allow_reserved_networks": self._allow_reserved_networks,
+            "concurrent": self._concurrent,
+            "limit_to_rfc1918": self._limit_to_rfc1918,
+            "filter": self._filter,
+            "filtermode": self._filter_mode,
+            "filename": self._filename,
+            "filemode": self._filemode,
+        }
+        return options_dict
+
     def items(self):
         """Return the IP address and hostnames as k, v pairs in list format."""
-        return {ip: hostname for ip, hostname in self._dns_dict.items()}.items()
+        return dict(self._dns_dict.items())
 
     def keys(self):
         """Return the IP address as keys in list format."""
-        return [ip for ip in self._dns_dict.keys()]
+        return list(self._dns_dict.keys())
 
-    def limit_to_rfc1918(self, value):
-        """Set the RFC1918 filter."""
-        if type(value) is not bool:
+    def limit_to_rfc1918(self, option):
+        """Set the RFC1918 filter.
+
+        Args:
+            option (bool): True limits the resolver to only IPv4 RFC1918 networks
+
+        IPv6 resolving is effecitvely disabled when this option is set to True
+        """
+        if not isinstance(option, bool):
             raise TypeError
-        self._limit_to_rfc1918 = value
+        self._limit_to_rfc1918 = option
 
     def loadfile(self, filename):
-        """Load saved data in YAML format."""
+        """Load saved data in YAML format.
+
+        Args
+            filename (str): path and filename for the disk based YAML cache file
+        """
         with open(filename) as f_handle:
             f_data = f_handle.read()
             self._dns_dict = yaml.safe_load(f_data)
 
     def remove_ip(self, ip):
-        """Remove an IP from the list, return false if not found."""
+        """Remove an IP from the list, return false if not found.
+
+        Args:
+            ip (str): A string containing an IP Address to remove
+        """
         try:
             IPAddress(ip)
             self._dns_dict.pop(ip)
-        except AddrFormatError:
-            raise TypeError
+        except AddrFormatError as error_case:
+            raise TypeError from error_case
         except KeyError:
             return False
         return True
 
     def _build_resolve_list(self):
         """Build list of IP's to perform resolver on, shared by serial and parallel methods."""
-        initial_pending_ips = [
+        if self._limit_to_rfc1918:
+            IPv4_skipped_networks = IPSet(IPV4_RESERVED_NETWORK_LIST)
+        elif self._allow_reserved_networks is False:
+            IPv4_skipped_networks = IPSet(IPV4_RESERVED_NETWORK_LIST)
+        else:
+            IPv4_skipped_networks = IPSet()
+
+        if self._allow_reserved_networks is False:
+            IPv6_skipped_networks = IPSet(IPV6_RESERVED_NETWORK_LIST)
+        else:
+            IPv6_skipped_networks = IPSet()
+
+        initial_ip_list = [
             key for key, value in self._dns_dict.items() if value is None
         ]
 
-        if self._limit_to_rfc1918:
-            IPv4_skipped_networks = IPSet(IPV4_RESERVED_NETWORK_LIST)
+        pending_ipset = IPSet(initial_ip_list)
+
+        if self._filter_mode == "block":
+            result_ipset = pending_ipset - self._filter
+            initial_pending_ips = [str(x) for x in result_ipset]
+        elif self._filter_mode == "allow":
+            result_ipset = pending_ipset & self._filter
+            initial_pending_ips = [str(x) for x in result_ipset]
         else:
-            IPv4_skipped_networks = IPSet(IPV4_RESERVED_NETWORK_LIST)
+            initial_pending_ips = [str(x) for x in pending_ipset]
 
-        IPv6_skipped_networks = IPSet(IPV6_RESERVED_NETWORK_LIST)
-
-        pending_ips = list()
+        pending_ips = []
 
         for key in initial_pending_ips:
             address = IPAddress(key)
@@ -295,7 +449,13 @@ class rdns_reaper:
             self._resolve_function(address)
 
     def savefile(self, filename=None):
-        """Save internal dictionary to YAML file."""
+        """Save internal dictionary to YAML file.
+
+        Args:
+            filename (str): Optional path and filename for disk based YAML cache file.
+                If not specified, the filename set at instance creation is used
+                If neither are specified, the operation will fail
+        """
         if filename is None and self._filename is not None and self._filemode == "w":
             filename = self._filename
 
@@ -303,13 +463,20 @@ class rdns_reaper:
             f_handle.write("---\n")
             f_handle.write(yaml.dump(self._dns_dict))
 
-    def setname(self, ip, hostname):
-        """Force the hostname of an IP."""
+    def setname(self, ip_address, hostname):
+        """Force the hostname of an IP.
+
+        Args:
+            ip_address (str): String containing an IP address to be modified
+            hostname (str): Desired FQDN hostname to be set for this entry
+                Can be None to reset record to allow for subsequent lookup
+        """
         try:
-            ip = IPAddress(ip)
-            self._dns_dict[ip] = hostname
-        except AddrFormatError:
-            raise TypeError
+            ip_address = IPAddress(ip_address)
+            self._dns_dict[ip_address] = hostname
+            return hostname
+        except AddrFormatError as error_case:
+            raise TypeError from error_case
         except KeyError:
             return False
 
@@ -318,12 +485,44 @@ class rdns_reaper:
         try:
             IPAddress(resolver_ip)
             self._resolver_ip = resolver_ip
-        except AddrFormatError:
+        except AddrFormatError as error_case:
+            raise TypeError from error_case
+
+    def set_filter(self, filter_input, **kwargs):
+        """Set customer filter options.
+
+        Args:
+            filter_input (str, list of strings, IPSet) - custom filter data
+                Data can be a single string containing an IP network, a list of strings
+                containing IP networks, or an IPSet containing networks
+
+            mode=("block" | "allow", optional) - named argument for a block list or allow list
+                a block list will allow anything not specified in the filter data
+                an allow list permits only what is specified in the filter data
+                the default option is a block list if no mode is specified
+        """
+        if kwargs.get("mode") is None:
+            self._filter_mode = "block"
+        elif kwargs.get("mode").lower() == "allow":
+            self._filter_mode = "allow"
+        elif kwargs.get("mode").lower() == "block":
+            self._filter_mode = "block"
+        else:
+            raise ValueError
+
+        if isinstance(filter_input, str):
+            filter_input = IPSet([filter_input])
+        elif isinstance(filter_input, list):
+            filter_input = IPSet(filter_input)
+
+        if not isinstance(filter_input, IPSet):
             raise TypeError
+
+        self._filter = filter_input
 
     def values(self):
         """Return the hostnames as values in list format."""
-        return [hostname for hostname in self._dns_dict.values()]
+        return list(self._dns_dict.values())
 
     @staticmethod
     def _isrfc1918(address_txt):
@@ -346,13 +545,14 @@ class rdns_reaper:
             if address_txt in reserved_network_IPSet:
                 return True
             return False
-        elif address.version == 6:
+
+        if address.version == 6:
             reserved_network_IPSet = IPSet(IPV6_RESERVED_NETWORK_LIST)
             if address_txt in reserved_network_IPSet:
                 return True
             return False
-        else:
-            raise ValueError
+
+        raise ValueError
 
     @staticmethod
     def _isreservedIPv6(address_txt):
@@ -364,8 +564,8 @@ class rdns_reaper:
             if address_txt in reserved_network_IPSet:
                 return True
             return False
-        else:
-            raise ValueError
+
+        raise ValueError
 
     class _reaper_iterator:
         def __init__(self, parentclass):
@@ -383,6 +583,5 @@ class rdns_reaper:
                 value = self.__parentclass._dns_dict[key]
                 self.__counter += 1
                 return (key, value)
-            else:
-                raise StopIteration
-            pass
+
+            raise StopIteration